backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
accelerate==0.29.3
    # via -r requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.9
    # via
    #   datasets
    #   fsspec
    #   langchain
    #   langchain-community
    #   realtime
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   langchain-core
    #   openai
    #   starlette
async-timeout==4.0.3
    # via
    #   -r requirements.in
    #   aiohttp
    #   langchain
attrs==25.3.0
    # via aiohttp
authlib==1.3.1
    # via weaviate-client
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
cryptography==45.0.3
    # via authlib
dataclasses-json==0.5.14
    # via
    #   langchain
    #   langchain-community
datasets==2.19.0
    # via -r requirements.in
deprecation==2.1.0
    # via
    #   postgrest
    #   weaviate-client
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
distro==1.9.0
    # via openai
exceptiongroup==1.3.0
    # via
    #   anyio
    #   pytest
fastapi==0.110.1
    # via -r requirements.in
filelock==3.18.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.6.2
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.3.1
    # via
    #   datasets
    #   huggingface-hub
    #   torch
gotrue==2.12.0
    # via supabase
greenlet==3.2.3
    # via sqlalchemy
grpcio==1.72.1
    # via
    #   grpcio-health-checking
    #   grpcio-tools
    #   weaviate-client
grpcio-health-checking==1.71.0
    # via weaviate-client
grpcio-tools==1.71.0
    # via weaviate-client
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via
    #   -r requirements.in
    #   gotrue
    #   openai
    #   postgrest
    #   storage3
    #   supabase
    #   supafunc
    #   weaviate-client
huggingface-hub==0.32.4
    # via
    #   accelerate
    #   datasets
    #   tokenizers
    #   transformers
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via torch
jiter==0.10.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.0.267
    # via -r requirements.in
langchain-community==0.0.10
    # via -r requirements.in
langchain-core==0.1.19
    # via
    #   -r requirements.in
    #   langchain-community
langsmith==0.0.83
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
mpmath==1.3.0
    # via sympy
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
mypy-extensions==1.1.0
    # via typing-inspect
networkx==3.4.2
    # via torch
numexpr==2.10.2
    # via langchain
numpy==1.26.4
    # via
    #   -r requirements.in
    #   accelerate
    #   datasets
    #   langchain
    #   langchain-community
    #   numexpr
    #   pandas
    #   transformers
openai==1.81.0
    # via -r requirements.in
openapi-schema-pydantic==1.2.4
    # via langchain
packaging==23.2
    # via
    #   -r requirements.in
    #   accelerate
    #   datasets
    #   deprecation
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
    #   pytest
    #   transformers
pandas==2.2.1
    # via
    #   -r requirements.in
    #   datasets
pluggy==1.6.0
    # via pytest
postgrest==1.0.2
    # via supabase
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.5
    # via
    #   -r requirements.in
    #   grpcio-health-checking
    #   grpcio-tools
psutil==7.0.0
    # via accelerate
psycopg2==2.9.10
    # via -r requirements.in
pyarrow==20.0.0
    # via datasets
pyarrow-hotfix==0.7
    # via datasets
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via
    #   -r requirements.in
    #   fastapi
    #   gotrue
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   openapi-schema-pydantic
    #   postgrest
    #   pydantic-settings
    #   weaviate-client
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via -r requirements.in
pygments==2.19.1
    # via pytest
pyjwt==2.10.1
    # via gotrue
pytest==8.4.0
    # via pytest-mock
pytest-mock==3.14.1
    # via gotrue
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   realtime
    #   storage3
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   transformers
realtime==2.4.3
    # via supabase
regex==2024.11.6
    # via
    #   tiktoken
    #   transformers
requests==2.32.3
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
    #   tiktoken
    #   transformers
safetensors==0.5.3
    # via
    #   accelerate
    #   transformers
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==1.4.54
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
starlette==0.37.2
    # via fastapi
storage3==0.11.3
    # via supabase
strenum==0.4.15
    # via
    #   postgrest
    #   supafunc
supabase==2.15.1
    # via -r requirements.in
supafunc==0.9.4
    # via supabase
sympy==1.14.0
    # via torch
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.5.2
    # via -r requirements.in
tokenizers==0.19.1
    # via transformers
tomli==2.2.1
    # via pytest
torch==2.7.0
    # via
    #   -r requirements.in
    #   accelerate
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   openai
    #   transformers
transformers==4.40.1
    # via -r requirements.in
typing-extensions==4.14.0
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   huggingface-hub
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   realtime
    #   torch
    #   typing-inspect
    #   typing-inspection
    #   uvicorn
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
validators==0.34.0
    # via weaviate-client
weaviate-client==4.15.0
    # via -r requirements.in
websockets==14.2
    # via realtime
xxhash==3.5.0
    # via datasets
yarl==1.20.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
